+++ docs/changes.rst
@@ -7,6 +7,7 @@ Not yet released.
 
 **Improvements**
 
+* :ref:`mt-google-translate-api-v3` now supports :ref:`glossary-mt` (optional).
 * A shortcut to duplicate a component is now available directly in the menu (:guilabel:`Manage` → :guilabel:`Duplicate Component`)
 * Included username when generating :ref:`credits`.
 * :ref:`bulk-edit` shows a preview of matched strings.